chore: moving `isForbiddenUrl` to `env.ts` for reuse
