+++ src/background/contentScriptHMR.ts
@@ -1,17 +1,4 @@
-import { isFirefox } from '~/env'
-
-const forbiddenProtocols = [
-  'chrome-extension://',
-  'chrome-search://',
-  'chrome://',
-  'devtools://',
-  'edge://',
-  'https://chrome.google.com/webstore',
-]
-
-function isForbiddenUrl(url: string): boolean {
-  return forbiddenProtocols.some(protocol => url.startsWith(protocol))
-}
+import { isFirefox, isForbiddenUrl } from '~/env'
 
 // Firefox fetch files from cache instead of reloading changes from disk,
 // hmr will not work as Chromium based browser
